# docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"HttpResponse_string_": {
				"properties": {
					"payload": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticateRequestBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TPost": {
				"properties": {
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"creator_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"content",
					"creator_id",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_TPost_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/TPost"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"USER_ROLES": {
				"enum": [
					"NORMAL",
					"ADMIN"
				],
				"type": "string"
			},
			"TokenPayload": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/USER_ROLES"
					}
				},
				"required": [
					"id",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePostRequestContent": {
				"properties": {
					"requestContent": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestContent",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeletePostRequest": {
				"properties": {
					"requestId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FetchPostsOutput": {
				"properties": {
					"creator": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"updatedAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"creator",
					"createdAt",
					"dislikes",
					"likes",
					"content",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_FetchPostsOutput-Array_": {
				"properties": {
					"payload": {
						"items": {
							"$ref": "#/components/schemas/FetchPostsOutput"
						},
						"type": "array"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LikeDislikePostRequest": {
				"properties": {
					"requestLike": {
						"type": "boolean"
					},
					"requestPostId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestLike",
					"requestPostId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.Exclude_keyofUser.password__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"role": {
						"$ref": "#/components/schemas/USER_ROLES"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"created_at",
					"role"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"HttpResponse_Omit_User.password__": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/Omit_User.password_"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequestBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Post": {
				"properties": {
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"creator_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"content",
					"creator_id",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_Post_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/Post"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePostRequestContent": {
				"properties": {
					"requestContent": {
						"type": "string"
					},
					"requestId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestContent",
					"requestId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "labook",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users/authenticate": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_string_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthenticateRequestBody"
							}
						}
					}
				}
			}
		},
		"/posts": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_TPost_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostRequestContent"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_FetchPostsOutput-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/posts/{id}": {
			"delete": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_string_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletePostRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_Post_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostRequestContent"
							}
						}
					}
				}
			}
		},
		"/posts/{id}/like": {
			"put": {
				"operationId": "Execute",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LikeDislikePostRequest"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_Omit_User.password__"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequestBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3003/v1"
		}
	]
}